append_to_destdir: handle Windows paths properly
It currently uses is_absolute() to strip one leading component which doesn't
work for paths like "C:/foo", "C:foo" and "/foo" which are either not absolute
or have multiple components that need to be stripped.

The goal is to join two paths and removing any components from the second
one that would result in the first one being replaced/changed.

push()/join() alters the first path if it starts with a RootDir, Prefix, or both, so
we push everything from the second path except for those.

Fixes #89