CI: add basics tests for Cygwin
Upstream Cygwin doesn't have a rust build yet (WIP) so use MSYS2
which ships rust for the Cygwin target.

For the example project don't run tests as those depend on inline-c
which depends on target-lexicon v0.11, while 0.13.3 is needed
for Cygwin support. /usr/local is not in the default pkgconf search
path, so add it.

Add separate jobs to avoid complicating the conditions of the existing
matrix setup even more.